+++ projects/generic/aim-tti-awg/src/egse/arbitrary_wave_generator/aim_tti/tgf4000_cs.py
@@ -194,18 +194,26 @@ def start(
 ) -> int:
     """Starts the Aim-TTi TGF4000 Control Server with the given identifier."""
 
-    # noinspection PyBroadException
-    try:
-        controller = Tgf4000ControlServer(device_id, simulator)
-        controller.serve()
-    except KeyboardInterrupt:
-        print("Shutdown requested...exiting")
-    except SystemExit as exc:
-        exit_code = exc.code if hasattr(exc, "code") else 0
-        print(f"System Exit with code {exc.code}")
-        sys.exit(exit_code)
-    except Exception:
-        logger.exception(f"Cannot start the Aim-TTi TGF4000 {device_id} Control Server")
+    multiprocessing.current_process().name = "daq6510_cs (start)"
+
+    with remote_logging():
+        from egse.env import setup_env
+
+        setup_env()
+
+        # noinspection PyBroadException
+        try:
+            control_server = Tgf4000ControlServer(device_id, simulator)
+            control_server.serve()
+        except KeyboardInterrupt:
+            logger.debug("Shutdown requested...exiting")
+        except SystemExit as exit_code:
+            logger.debug("System Exit with code {}.".format(exit_code))
+            sys.exit(exit_code.code)
+        except Exception:
+            msg = "Cannot start the Aim-TTi TGF4000 {device_id} Control Server"
+            logger.exception(msg)
+            rich.print(f"[red]{msg}.")
 
     return 0
 
@@ -297,7 +305,7 @@ def status(
 if __name__ == "__main__":
     import logging
 
-    from egse.logger import set_all_logger_levels
+    from egse.logger import set_all_logger_levels, remote_logging
 
     set_all_logger_levels(logging.DEBUG)